Change performance points data type to float

+++ BanchoMultiplayerBot.Osu/Data/BeatmapPerformanceInfo.cs
@@ -6,11 +6,11 @@ namespace BanchoMultiplayerBot.Osu.Data;
 public class BeatmapPerformanceInfo
 {
     [JsonPropertyName("pp100")]
-    public int Performance100 { get; init; }
+    public float Performance100 { get; init; }
     
     [JsonPropertyName("pp98")]
-    public int Performance98 { get; init; }
+    public float Performance98 { get; init; }
     
     [JsonPropertyName("pp95")]
-    public int Performance95 { get; init; }
+    public float Performance95 { get; init; }
 }
